+++ crates/sml-statics/src/dec.rs
@@ -2,7 +2,7 @@
 
 use crate::env::{Cx, Env, EnvLike as _};
 use crate::error::{ErrorKind, Item};
-use crate::generalize::{generalize, generalize_fixed, HasRecordMetaVars};
+use crate::generalize::{generalize, generalize_fixed, HasRecordMetaVar};
 use crate::get_env::{get_env_from_str_path, get_ty_info, get_val_info};
 use crate::types::{
   Equality, FixedTyVars, Generalizable, IdStatus, StartedSym, Ty, TyEnv, TyInfo, TyScheme,
@@ -100,8 +100,8 @@ pub(crate) fn get(
             ErrorKind::BindPolymorphicExpansiveExp,
           );
         }
-        if let Err(HasRecordMetaVars) = g {
-          st.err(dec, ErrorKind::UnresolvedRecordTy);
+        if let Err(HasRecordMetaVar(idx)) = g {
+          st.err(idx, ErrorKind::UnresolvedRecordTy);
         }
       }
       // extend the overall env with that.